requirements.txt
python-dotenv
requests
pandas
tqdm
together
matplotlib
seaborn

